refactor - Added user not found error handling in the user base service.
