+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/service/user/RandomNameService.java
@@ -1,4 +1,4 @@
-package dev.anirban.harmoniq_backend.service;
+package dev.anirban.harmoniq_backend.service.user;
 
 import org.springframework.stereotype.Service;
